pizzalibrary/pyproject.toml
[tool.poetry]
name = "pizzalibrary"
version = "0.1.0"
description = ""
authors = ["Andre Bossard <anbossar@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.2"
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
